feat: remove overview on sidebar
