--- apps/dashboard/src/routes/courses/[id]/newsfeed/+page.js
@@ -1,5 +1,0 @@
-export const load = ({ params = { id: '' } }) => {
-  return {
-    courseId: params.id
-  };
-};
